Prevent calling callback twice
Preact registers on elements event listeners for properties starting with 'on'.
WC are also just html elements. When we observe e.g. onClick attribute on our wc
we don't want preact to register native 'click' event for it. Instead we want it to be
passed as a prop.
This commit override the event handler which preact registered.